Revisions for all classes (see details)
Went through the code and tried to simplify what I could. Also noticed a few array lists were not initialized and some to Strings were clunky.

CorornaPassport:
Cleaned up the constructor.

Course:
Schedule was not initialized in the constructor. Also added the course name in the toString.

Date:
Changed the toString to match the requirements in the note for lesson.

DiscordLesson:
Added a channel string before the channel.

Grade:
Added the course to constructor where it was needed. Also changed the toString to not print all the damn lessons of the course.

Lesson:
According the class diagram, the resources can be empty, so I commented out the exception for the resources being empty. In addition, the times must be valid for a lesson to be created. That method was also made static to reflect the class diagram.
I changed the getDuration to use the timeUntil method from time.
The getDateTimeString was changed to reflect the note on lesson. It also uses the toString from Date.
The equals method was always returning false because the comparison on the arraylists was malformed. First, the pdfs are part of the resources, so that was removed. Then a boolean that checks each resource against the other, given the sizes are the same, was added.
Lastly, the toString was changed to list the resources in a pleasant way at the end.

OnlineLesson
Changed the toString to make it easier to read.

Resource:
isPDF was modified to accept "PDF" as an extension, given the note on the diagram.
ToString was modified to be shorter and appear cleaner in the lesson toString.

Student:
Added todo to check all of the classes to make sure all of the methods are present because getCoronaPassport was not present.
In addGrade, an exception is needed for when the course does not match one a student has.
Corona passport becomes invalid for a student when there are more than 180 days after the infection, so I changed the comparator to match.

Time:
Made the constructor check to see the time created is valid.
Ran the formatter on the class.

ZoomLesson:
Added a colon and a space to the toString.